# apps/backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/quizzes": {
      "get": {
        "operationId": "QuizzesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuizDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      },
      "post": {
        "operationId": "QuizzesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/samples": {
      "post": {
        "deprecated": true,
        "description": "Kept for existing integrations. The samples now live in the template catalogue, copied on demand.",
        "operationId": "QuizzesController_createSamples",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuizDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Adds the sample quizzes to the caller's bank",
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/import": {
      "post": {
        "operationId": "QuizzesController_importQuiz",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/export": {
      "get": {
        "operationId": "QuizzesController_exportQuiz",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Zip bundle (quiz.json + media/)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/publication": {
      "get": {
        "operationId": "QuizzesController_publicationReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationReportDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/publication/export": {
      "post": {
        "operationId": "QuizzesController_publicationExport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationExportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Zip bundle (quiz.json + media/), named after the slug."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/duplicate": {
      "post": {
        "operationId": "QuizzesController_duplicate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}": {
      "get": {
        "operationId": "QuizzesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      },
      "put": {
        "operationId": "QuizzesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuizDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      },
      "delete": {
        "operationId": "QuizzesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/feedback": {
      "get": {
        "operationId": "QuizzesController_feedback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "rating",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 5,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizFeedbackSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/sessions": {
      "get": {
        "operationId": "QuizzesController_sessions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/sessions/{sessionId}": {
      "get": {
        "operationId": "QuizzesController_sessionDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/sessions/{sessionId}/players/{playerResultId}": {
      "get": {
        "operationId": "QuizzesController_sessionPlayer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "playerResultId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionPlayerDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/status": {
      "patch": {
        "operationId": "QuizzesController_transition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionQuizDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/media": {
      "get": {
        "operationId": "MediaController_list",
        "parameters": [
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "image",
                "video",
                "audio"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaLibraryItemDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "media"
        ]
      },
      "post": {
        "operationId": "MediaController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "durationMs": {
                    "type": "integer"
                  },
                  "peaks": {
                    "type": "string",
                    "description": "JSON array of AUDIO_PEAK_COUNT values in 0–1."
                  },
                  "origin": {
                    "type": "string",
                    "enum": [
                      "upload",
                      "recording"
                    ]
                  },
                  "loudnessLufs": {
                    "type": "number"
                  },
                  "peakDbfs": {
                    "type": "number"
                  },
                  "sourceSha256": {
                    "type": "string",
                    "description": "SHA-256 of the original file, before conversion."
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaUploadResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "media"
        ]
      }
    },
...
    "/api/v1/media/{id}": {
      "get": {
        "operationId": "MediaController_serve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contenu binaire du média."
          },
          "206": {
            "description": "Partie demandée par `Range`."
          },
          "416": {
            "description": "Plage hors du fichier."
          }
        },
        "tags": [
          "media"
        ]
      },
      "delete": {
        "operationId": "MediaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "409": {
            "description": "Still used by a quiz or a session."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "media"
        ]
      }
    },
...
  },
  "info": {
    "title": "QuizDock API",
    "description": "API REST du builder de quiz et des restitutions",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "QuizDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "coverMediaId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "ready",
              "archived"
            ]
          },
          "language": {
            "type": "string"
          },
          "feedbackEnabled": {
            "type": "boolean"
          },
          "mediaTailS": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "loudnessTargetLufs": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "audioTarget": {
            "type": "string",
            "enum": [
              "projection",
              "projection_remote",
              "everyone"
            ]
          },
          "questionCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "license": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "shared": {
            "type": "boolean"
          },
          "editable": {
            "type": "boolean"
          },
          "ownerName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "archivedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "ownerId",
          "title",
          "description",
          "coverMediaId",
          "status",
          "language",
          "feedbackEnabled",
          "mediaTailS",
          "loudnessTargetLufs",
          "audioTarget",
          "questionCount",
          "license",
          "tags",
          "shared",
          "createdAt",
          "updatedAt",
          "archivedAt"
        ]
      },
      "CreateQuizDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "language": {
            "type": "string",
            "maxLength": 10,
            "pattern": "^[a-z]{2,3}(-[A-Za-z0-9]{2,8})*$"
          },
          "feedbackEnabled": {
            "type": "boolean"
          },
          "mediaTailS": {
            "type": "integer",
            "minimum": 0,
            "maximum": 30
          },
          "loudnessTargetLufs": {
            "anyOf": [
              {
                "type": "number",
                "enum": [
                  -14
                ]
              },
              {
                "type": "number",
                "enum": [
                  -16
                ]
              },
              {
                "type": "number",
                "enum": [
                  -23
                ]
              }
            ]
          },
          "audioTarget": {
            "type": "string",
            "enum": [
              "projection",
              "projection_remote",
              "everyone"
            ]
          },
          "coverMediaId": {
            "type": "string",
            "minLength": 26,
            "maxLength": 26
          }
        },
        "required": [
          "title"
        ]
      },
      "PublicationReportDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "slugSet": {
            "type": "boolean"
          },
          "language": {
            "type": "string"
          },
          "license": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "estimatedBytes": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "maxBytes": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "enum": [
                    "not_ready",
                    "license",
                    "language",
                    "tags",
                    "too_large",
                    "credit_missing"
                  ]
                },
                "level": {
                  "type": "string",
                  "enum": [
                    "block",
                    "warn"
                  ]
                },
                "count": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "code",
                "level"
              ]
            }
          },
          "heaviest": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "kind": {
                  "type": "string",
                  "enum": [
                    "image",
                    "video",
                    "audio"
                  ]
                },
                "name": {
                  "type": "string",
                  "nullable": true
                },
                "sizeBytes": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "id",
                "kind",
                "name",
                "sizeBytes"
              ]
            }
          },
          "uncredited": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "kind": {
                  "type": "string",
                  "enum": [
                    "image",
                    "video",
                    "audio"
                  ]
                },
                "name": {
                  "type": "string",
                  "nullable": true
                },
                "sizeBytes": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "id",
                "kind",
                "name",
                "sizeBytes"
              ]
            }
          }
        },
        "required": [
          "slug",
          "slugSet",
          "language",
          "license",
          "tags",
          "estimatedBytes",
          "maxBytes",
          "issues",
          "heaviest",
          "uncredited"
        ]
      },
      "PublicationExportDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 60
          }
        },
        "required": [
          "slug"
        ]
      },
...
      "UpdateQuizDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          },
          "language": {
            "type": "string",
            "maxLength": 10,
            "pattern": "^[a-z]{2,3}(-[A-Za-z0-9]{2,8})*$"
          },
          "feedbackEnabled": {
            "type": "boolean"
          },
          "mediaTailS": {
            "type": "integer",
            "minimum": 0,
            "maximum": 30
          },
          "loudnessTargetLufs": {
            "anyOf": [
              {
                "type": "number",
                "enum": [
                  -14
                ]
              },
              {
                "type": "number",
                "enum": [
                  -16
                ]
              },
              {
                "type": "number",
                "enum": [
                  -23
                ]
              }
            ]
          },
          "audioTarget": {
            "type": "string",
            "enum": [
              "projection",
              "projection_remote",
              "everyone"
            ]
          },
          "coverMediaId": {
            "type": "string",
            "minLength": 26,
            "maxLength": 26,
            "nullable": true
          },
          "license": {
            "type": "string",
            "enum": [
              "CC0-1.0",
              "CC-BY-4.0",
              "CC-BY-SA-4.0",
              null
            ],
            "nullable": true
          },
          "shared": {
            "type": "boolean"
          },
          "tags": {
            "maxItems": 5,
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 30,
              "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
            }
          }
        }
      },
...
    }
  }
}